feat: 워크스페이스 crud, 멤버 초대, 초대 상태 수정

+++ src/main/java/onepick/kanban/workspace/service/WorkspaceService.java
@@ -3,6 +3,9 @@ package onepick.kanban.workspace.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import onepick.kanban.common.SlackNotifier;
+import onepick.kanban.user.entity.Role;
+import onepick.kanban.user.entity.User;
+import onepick.kanban.user.repository.UserRepository;
 import onepick.kanban.workspace.dto.WorkspaceRequestDto;
 import onepick.kanban.workspace.dto.WorkspaceResponseDto;
 import onepick.kanban.workspace.entity.Workspace;
@@ -18,6 +21,7 @@ public class WorkspaceService {
 
     private final WorkspaceRepository workspaceRepository;
     private final SlackNotifier slackNotifier;
+    private final UserRepository userRepository;
 
     @Transactional
     public WorkspaceResponseDto createWorkspace(WorkspaceRequestDto requestDto) {
@@ -47,9 +51,16 @@ public class WorkspaceService {
     }
 
     @Transactional
-    public void deleteWorkspace(Long workspaceId) {
+    public void deleteWorkspace(Long workspaceId, Long deleterId) {
         Workspace workspace = workspaceRepository.findById(workspaceId)
                 .orElseThrow(() -> new IllegalArgumentException("워크스페이스를 찾을 수 없습니다."));
+
+        User deleter = userRepository.findById(deleterId)
+                .orElseThrow(() -> new IllegalArgumentException("삭제 요청 사용자를 찾을 수 없습니다."));
+
+        if (!deleter.getRole().equals(Role.WORKSPACE_ADMIN) && !deleter.getRole().equals(Role.ADMIN)) {
+            throw new IllegalArgumentException("워크스페이스 삭제 권한이 없는 사용자입니다.");
+        }
         workspaceRepository.delete(workspace);
     }
 }
